requirements.txt
nltk
progressbar2
beautifulsoup4
python-crfsuite
numpy
scikit_learn